+++ src/views/Hub/HubHome.vue
@@ -88,12 +88,14 @@
                             v-if="isItNew(article.date)"
                             label
                             small
-                            color="#0D4474"
                             class="mr-2"
                             style="margin-top: 0px"
                           >
                             <span
-                              style="color: #fff !important; font-weight: 400"
+                              style="
+                                color: #000000 !important;
+                                font-weight: 700;
+                              "
                             >
                               NEW!
                             </span>